+++ ports/libsimpleservo/api/src/lib.rs
@@ -185,7 +185,12 @@ pub struct ServoGlue {
 }
 
 pub fn servo_version() -> String {
-    servo::config::servo_version()
+    let cargo_version = env!("CARGO_PKG_VERSION");
+    let git_info = option_env!("GIT_INFO");
+    match git_info {
+        Some(info) => format!("Servo {}{}", cargo_version, info),
+        None => format!("Servo {}", cargo_version),
+    }
 }
 
 /// Test if a url is valid.
